fix: usar Decimal/Numeric en vez de float para montos de dinero (#4)
Los montos (total, cobrado, saldo, costo, cargo, abono, etc.) ahora se
manejan como Decimal en Python y se guardan en columnas Numeric(10, 2)
en SQLite, en vez de float/Float. float no puede representar
exactamente la mayoria de los montos decimales, y sumar o restar
dinero repetidamente (como pasa en cada pago, anulacion o devolucion)
puede acumular errores de redondeo con el tiempo. Decimal evita ese
problema de raiz en vez de depender de round(x, 2) despues de cada
operacion.

Cambios:
- Nuevo helper redondear_monto() en services/comun.py: convierte
  cualquier valor a Decimal con 2 decimales exactos.
- models/finanzas.py y models/cita.py: columnas de dinero de Float a
  Numeric(10, 2).
- schemas/finanzas.py y schemas/cita.py: campos de dinero (monto,
  costo, montoPagado) de float a Decimal en los payloads Pydantic.
- services/finanzas.py y services/citas.py: toda la aritmetica de
  dinero reescrita con Decimal en vez de round(float(x), 2).
- Nuevo tests/test_precision_monetaria.py: prueba de regresion que
  llama la funcion real registrar_pago() 500 veces seguidas y confirma
  que el saldo final es exacto.

No requiere migracion de base de datos: SQLite no impone tipos de
columna de forma estricta, asi que las bases de datos existentes (con
la columna declarada como FLOAT) siguen funcionando sin cambios; los
valores viejos se siguen leyendo bien y todo lo que se escriba de aca
en adelante queda exacto. Verificado con una base de datos de prueba
creada con el esquema anterior.

70 tests existentes + 2 nuevos pasan. Lint y format (ruff) limpios.

+++ backend/app/models/cita.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Column, Float, Integer, String, Text
+from sqlalchemy import JSON, Column, Integer, Numeric, String, Text
 
 from ..database import Base
 
@@ -18,7 +18,7 @@ class CitaDB(Base):
     servicios = Column(JSON)
     notas = Column(Text)
     notasFin = Column(Text)
-    costo = Column(Float)
+    costo = Column(Numeric(10, 2))
     tipoPago = Column(String(50))
     estado = Column(String(50), index=True)
     sesionNum = Column(Integer)
